requirements.txt
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
python-dotenv==1.1.1
dashscope==1.24.4
opencc-python-reimplemented==0.1.7
streamlit==1.49.1